+++ src/terminal/getcmd.c
@@ -6,7 +6,7 @@
 /*   By: ciglesia <ciglesia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/17 19:17:23 by ciglesia          #+#    #+#             */
-/*   Updated: 2021/07/30 19:07:39 by ciglesia         ###   ########.fr       */
+/*   Updated: 2021/08/01 01:55:56 by user             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
