+++ src/terminal/line_editor/move_cursor.c
@@ -1,50 +1,34 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   multiline.c                                        :+:      :+:    :+:   */
+/*   move_cursor.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ciglesia <ciglesia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/07/29 00:42:57 by ciglesia          #+#    #+#             */
-/*   Updated: 2021/07/30 21:42:57 by ciglesia         ###   ########.fr       */
+/*   Created: 2021/07/30 16:23:15 by ciglesia          #+#    #+#             */
+/*   Updated: 2021/07/31 22:35:27 by ciglesia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "msh.h"
 
-t_coords	cursor_position(void)
+//int	reset_cursor(){}
+
+int	move_arrows(char *buf)
 {
-	t_coords	c;
-	long		size;
-	long		cursor;
-
-	ioctl(STDIN_FILENO, TIOCGWINSZ, &g_sh->events->ws);
-	c.len = g_sh->events->ws.ws_col;
-	size = g_sh->line_size;
-	cursor = g_sh->line_cursor;
-	c.ll = 0;
-	if (size > (c.len - PROMPT_LEN))
+	reset_shadow();
+	if (!ft_strcmp(g_sh->events->lf, buf) && g_sh->line_cursor)
 	{
-		size -= (c.len - PROMPT_LEN);
-		c.ll = 1;
+		g_sh->line_cursor--;
+		return (2);
 	}
-	c.cl = 0;
-	if (cursor > (c.len - PROMPT_LEN))
+	if (!ft_strcmp(g_sh->events->rg, buf) && g_sh->line_cursor
+		< ft_strlen(g_sh->line))
 	{
-		cursor -= (c.len - PROMPT_LEN);
-		c.cl = 1;
+		g_sh->line_cursor++;
+		return (2);
 	}
-
-	c.lc = size % c.len;
-	if (!c.ll)
-		c.lc = size % (c.len - PROMPT_LEN);
-	c.ll += (size / c.len) + (c.lc != 0);
-
-	c.cc = cursor % c.len;
-	if (!c.cl)
-		c.cc = cursor % (c.len - PROMPT_LEN);
-	c.cl += (cursor / c.len) + (c.cc != 0);
-	return (c);
+	return (0);
 }
 
 /*
@@ -80,48 +64,6 @@ void	move_vertically(char *buf)
 	}
 }
 
-static int	next_space(char *str, int i)
-{
-	int	my_space;
-
-	if (str && str[i - 1])
-	{
-		my_space = 0;
-		i--;
-		if (str[i] == ' ')
-			my_space = 1;
-		while (i && (str[i] != ' ' || my_space))
-		{
-			if (str[i] != ' ')
-				my_space = 0;
-			i--;
-		}
-		if (!my_space && str[i] == ' ')
-			i++;
-	}
-	return (i);
-}
-
-static int	next_char(char *str, int i)
-{
-	int	my_char;
-
-	if (str && str[i])
-	{
-		my_char = 0;
-		i++;
-		if (str[i] != ' ')
-			my_char = 1;
-		while (str[i] && (str[i] == ' ' || my_char))
-		{
-			if (str[i] == ' ')
-				my_char = 0;
-			i++;
-		}
-	}
-	return (i);
-}
-
 int	move_ctrl(char *buf)
 {
 	long	x;
@@ -147,5 +89,26 @@ int	move_ctrl(char *buf)
 	}
 	else
 		move_vertically(buf);
+	return (0);
+}
+
+int	jump_sides(char *buf)
+{
+	if (buf[2] == 70 && g_sh->line_cursor < ft_strlen(g_sh->line))
+	{
+		while (g_sh->line_cursor < ft_strlen(g_sh->line))
+		{
+			ft_putstr_fd(g_sh->events->rg, 0);
+			g_sh->line_cursor++;
+		}
+	}
+	if (buf[2] == 72 && g_sh->line_cursor)
+	{
+		while (g_sh->line_cursor)
+		{
+			ft_putstr_fd(g_sh->events->lf, 0);
+			g_sh->line_cursor--;
+		}
+	}
 	return (0);
 }
